+++ src/controllers/AgenteController.ts
@@ -1,7 +1,6 @@
 import { Request,Response } from "express";
 import AbstractController from "./AbstractController";
 import connect from "../services/connectService";
-import AWS from "../services/amazonSNS";
 
 
 class AgenteController extends AbstractController{
@@ -168,32 +167,6 @@ class AgenteController extends AbstractController{
             res.status(500).send('Internal server error' + err);
         }
     }
-
-    private async postEnviarSMS(req:Request, res:Response){
-        try{
-            const service = req.body.service
-            const name = req.body.clientName
-            const direccion = req.body.direccion
-            const params = {
-                Message: name + ", tu " + service + " va en camino a: " + direccion,
-                PhoneNumber: req.body.number,
-                MessageAttributes: {
-                        'AWS.SNS.SMS.SenderID': {
-                            'DataType': 'String',
-                            'StringValue': 'String'
-                    }
-                }
-            };
-            const mensaje = await new AWS.SNS().publish(params).promise();
-            console.log(mensaje);
-            res.status(200).send("<h1>Mensaje mandado</h1>");
-        } catch(err){
-            console.log(err);
-            res.status(500).send('Internal server error '+ err);
-        }
-
-    }
-
 
 }
